Refactor HKLinfo program into crys3d/programs/HKLinfo.py and crys3d/command_line/HKLinfo.py. Add crys3d test to azure_pipelines.
